refactor(logging): address review comments
see #127
